Move fixtures to conftest.
